bot/requirements.txt
python-telegram-bot
python-dotenv~=1.0.1
requests